firefox: use the link_regex from the preferences dialogue in filtering
This turned out to be much messier than I anticipated as we are trying
to do something with a context menu that is not usually done in
this way.  Have no fear, the hacks are here!  We make it work by
passing the link_regex through via a content script and removing/adding
the context menu item whenever anything changes.  Lovely? no. Functinoal?
Yes.

Signed-off-by: Paul Osborne <osbpau@gmail.com>